Updated 'rebalancing_dates' variable
Renamed it as REBALANCING_DATES. Moreover, the previous way to determine the rebalancing dates, which is to use business dates, was wrong. For example, 2012-04-30 is a business date but not a trading date in the Chinese market. This is fixed by using the last date of each month in 'df_basic_info', which contains all trading dates, as the rebalancing dates.